Excel to Dayabase done

+++ Models/SaleModel.cs
@@ -12,7 +12,7 @@ namespace ExcelReporting.Models
         public string? Country { get; set; }
         public string? Product { get; set; }
         public Discount DiscountBand  { get; set; }
-        public int UnitsSold { get; set; }
+        public decimal UnitsSold { get; set; }
         public decimal ManufacturingPrice { get; set; }
         public decimal SalePrice { get; set; }
         public decimal GrossSales { get; set; }
